// src/app/models/classes/detail-invoice.ts
export class DetailInvoice {
    actividadEconomica: number;
    codigoProductoSin: number;
    codigoProducto: string;
    descripcion: string;
    cantidad: number;
    unidadMedida: string;
    precioUnitario: string;
    montoDescuento: number;
    subTotal: number;
    numeroSerie: number;
    numeroImei: number;
}